Move `OplCapture` into `opl_capture.cpp`
