+++ src/hardware/meson.build
@@ -44,6 +44,7 @@ libhardware_sources = files(
     'mpu401.cpp',
     'ne2000.cpp',
     'opl.cpp',
+    'opl_capture.cpp',
     'pci_bus.cpp',
     'pcspeaker.cpp',
     'pcspeaker_discrete.cpp',
