+++ vs/dosbox.vcxproj
@@ -617,6 +617,7 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\hardware\mpu401.cpp" />
     <ClCompile Include="..\src\hardware\ne2000.cpp" />
     <ClCompile Include="..\src\hardware\opl.cpp" />
+    <ClCompile Include="..\src\hardware\opl_capture.cpp" />
     <ClCompile Include="..\src\hardware\pci_bus.cpp" />
     <ClCompile Include="..\src\hardware\pcspeaker.cpp" />
     <ClCompile Include="..\src\hardware\pcspeaker_discrete.cpp" />
@@ -889,6 +890,7 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\src\hardware\innovation.h" />
     <ClInclude Include="..\src\hardware\lpt_dac.h" />
     <ClInclude Include="..\src\hardware\opl.h" />
+    <ClInclude Include="..\src\hardware\opl_capture.h" />
     <ClInclude Include="..\src\hardware\pcspeaker.h" />
     <ClInclude Include="..\src\hardware\pcspeaker_discrete.h" />
     <ClInclude Include="..\src\hardware\pcspeaker_impulse.h" />
