+++ vs/dosbox.vcxproj.filters
@@ -268,6 +268,9 @@
     <ClCompile Include="..\src\hardware\opl.cpp">
       <Filter>src\hardware</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\hardware\opl_capture.cpp">
+      <Filter>src\hardware</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\hardware\pci_bus.cpp">
       <Filter>src\hardware</Filter>
     </ClCompile>
@@ -1255,6 +1258,9 @@
       <Filter>src\hardware</Filter>
     </ClInclude>
     <ClInclude Include="..\src\hardware\opl.h">
+      <Filter>src\hardware</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\hardware\opl_capture.h">
       <Filter>src\hardware</Filter>
     </ClInclude>
     <ClInclude Include="..\src\hardware\input\intel8042.h">
